requirements.txt
fsspec==2024.12.0
huggingface_hub>=0.24
numba-cuda==0.15.1
cuda-bindings
numexpr<2.14.0
numpy>=1.22
onnx>=1.7.0
protobuf~=5.29.5
python-dateutil
ruamel.yaml
scikit-learn
setuptools>=70.0.0
tensorboard
text-unidecode
torch>=2.6.0
torchaudio>=2.6.0
tqdm>=4.41.0
wget
wrapt
lightning>2.2.1,<=2.4.0
soundfile>=0.12
datasets>=2.14
transformers
maai>=0.1.0
scipy>=1.10
gliner>=0.2
openai>=1.0
httpx>=0.27
supermemory>=3.4
python-dotenv>=1.0
websockets>=12
elevenlabs>=1.0
